refactor(nextcloud): Use FolderBody
